修复查看大图时崩溃的问题，优化应用启动速度，删除无用导包，优化代码逻辑
