+++ app/src/main/java/com/hjq/demo/ui/activity/PersonalDataActivity.java
@@ -17,7 +17,7 @@ import butterknife.OnClick;
  *    author : Android 轮子哥
  *    github : https://github.com/getActivity/AndroidProject
  *    time   : 2019/04/20
- *    desc   : 个人资料界面
+ *    desc   : 个人资料
  */
 public final class PersonalDataActivity extends MyActivity {
