+++ app/src/main/res/layout/fragment_test_b.xml
@@ -33,13 +33,13 @@
                 <de.hdodenhof.circleimageview.CircleImageView
                     android:layout_width="@dimen/space_180"
                     android:layout_height="@dimen/space_180"
-                    android:src="@mipmap/bg_launcher" />
+                    android:src="@drawable/bg_launcher" />
 
                 <de.hdodenhof.circleimageview.CircleImageView
                     android:layout_width="@dimen/space_180"
                     android:layout_height="@dimen/space_180"
                     android:layout_marginLeft="@dimen/space_40"
-                    android:src="@mipmap/ic_logo"
+                    android:src="@drawable/ic_logo"
                     app:civ_border_color="@color/white60"
                     app:civ_border_width="@dimen/space_2" />
 
@@ -47,7 +47,7 @@
                     android:layout_width="@dimen/space_180"
                     android:layout_height="@dimen/space_180"
                     android:layout_marginLeft="@dimen/space_40"
-                    android:src="@mipmap/ic_logo"
+                    android:src="@drawable/ic_logo"
                     app:civ_border_color="@color/black60"
                     app:civ_border_width="@dimen/space_2" />
 
@@ -210,7 +210,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:scaleType="centerCrop"
-                android:src="@mipmap/bg_launcher" />
+                android:src="@drawable/bg_launcher" />
 
         </LinearLayout>
