--- umeng/src/main/java/com/hjq/umeng/WeChatActivity.java
@@ -1,15 +1,0 @@
-package com.hjq.umeng;
-
-import android.os.Bundle;
-
-import com.umeng.socialize.weixin.view.WXCallbackActivity;
-
-/**
- *    author : Android 轮子哥
- *    github : https://github.com/getActivity/AndroidProject
- *    time   : 2019/04/29
- *    desc   : 微信回调 Activity
- */
-public class WeChatActivity extends WXCallbackActivity {
-
-}
